backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27

pydantic>=2.6,<3
pydantic-settings>=2.2,<3
email-validator>=2.1.0

sqlalchemy>=2.0
psycopg[binary]>=3.1        # psycopg v3 (rimuove psycopg2)
alembic>=1.13

fredapi>=0.5
pandas>=2.0
numpy>=1.26
yfinance>=0.2
python-dateutil>=2.8
apscheduler>=3.10

python-jose[cryptography]>=3.3
passlib[bcrypt]
bcrypt==3.2.2

httpx>=0.27
pytest>=8.0